+++ apps/backoffice-v2/src/common/utils/check-is-individual/check-is-individual.ts
@@ -1,0 +1,4 @@
+import { TWorkflowById } from '@/domains/workflows/fetchers';
+
+export const checkIsIndividual = (workflow: TWorkflowById) =>
+  workflow?.context?.entity?.type === 'individual';