+++ apps/kyb-app/DynamicElements/hooks/useDynamicUIContext/useDynamicUIContext.ts
@@ -1,8 +1,0 @@
-import {
-  DynamicUIRendererContext,
-  dynamicUIRendererContext,
-} from '@/components/organisms/DynamicElements/context';
-import { useContext } from 'react';
-
-export const useDynamicUIContext = <TContext>() =>
-  useContext(dynamicUIRendererContext) as DynamicUIRendererContext<TContext>;