+++ apps/kyb-app/src/common/components/layouts/Signup/context/SignupLayoutProvider/signup-layout.context.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import { ISignupLayoutContext } from './types';
+
+export const SignupLayoutContext = createContext<ISignupLayoutContext>({});