+++ apps/kyb-app/src/common/providers/AccessTokenProvider/context/access-token-context.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+import { IAccessTokenContext } from './types';
+
+export const AccessTokenContext = createContext<IAccessTokenContext>({
+  accessToken: null,
+  setAccessToken: () => {},
+});