+++ packages/common/src/utils/safe-every/safe-every.ts
@@ -1,4 +1,7 @@
-export const safeEvery = <TItem>(array: Array<TItem>, predicate: (item: TItem) => boolean) => {
+export const safeEvery = <TItem>(
+  array: TItem[] | readonly TItem[],
+  predicate: (item: TItem) => boolean,
+) => {
   if (!Array.isArray(array) || !array?.length) return false;
 
   return array.every(predicate);