+++ services/workflows-service/src/workflow/workflow.controller.external.ts
@@ -6,7 +6,6 @@ import { HttpStatus, NotFoundException, Query, Res } from '@nestjs/common';
 import * as swagger from '@nestjs/swagger';
 import { ApiOkResponse, ApiResponse } from '@nestjs/swagger';
 import type { WorkflowRuntimeData } from '@prisma/client';
-// import * as nestAccessControl from 'nest-access-control';
 import { WorkflowTokenService } from '@/auth/workflow-token/workflow-token.service';
 import { putPluginsExampleResponse } from '@/workflow/workflow-controller-examples';
 import { CurrentProject } from '@/common/decorators/current-project.decorator';
@@ -19,7 +18,7 @@ import { PrismaService } from '@/prisma/prisma.service';
 import type { AnyRecord, InputJsonValue, TProjectId, TProjectIds } from '@/types';
 import { WORKFLOW_DEFINITION_TAG } from '@/workflow-defintion/workflow-definition.controller';
 import { WorkflowDefinitionService } from '@/workflow-defintion/workflow-definition.service';
-import { CreateCollectionFlowUrlDto } from '@/workflow/dtos/create-collection-flow-url';
+import { CreateCollectionFlowUrlDto } from '@/workflow/dtos/create-collection-flow-url.dto';
 import { GetWorkflowsRuntimeInputDto } from '@/workflow/dtos/get-workflows-runtime-input.dto';
 import { GetWorkflowsRuntimeOutputDto } from '@/workflow/dtos/get-workflows-runtime-output.dto';
 import { WorkflowHookQuery } from '@/workflow/dtos/workflow-hook-query';
@@ -45,6 +44,7 @@ import { defaultContextSchema } from '@ballerine/common';
 import { WorkflowRunSchema } from './schemas/workflow-run';
 import { ValidationError } from '@/errors';
 import { WorkflowRuntimeListItemModel } from '@/workflow/workflow-runtime-list-item.model';
+import { CreateTokenDto } from '@/workflow/dtos/create-token.dto';
 
 export const WORKFLOW_TAG = 'Workflows';
 @swagger.ApiBearerAuth()
@@ -54,8 +54,6 @@ export class WorkflowControllerExternal {
   constructor(
     protected readonly service: WorkflowService,
     protected readonly normalizeService: HookCallbackHandlerService,
-    // @nestAccessControl.InjectRolesBuilder()
-    // protected readonly rolesBuilder: nestAccessControl.RolesBuilder,
     private readonly workflowTokenService: WorkflowTokenService,
     private readonly workflowDefinitionService: WorkflowDefinitionService,
     private readonly prismaService: PrismaService,
@@ -146,7 +144,6 @@ export class WorkflowControllerExternal {
       },
       {
         type: 'body',
-
         schema: WorkflowExtensionSchema,
       },
     ],
@@ -381,20 +378,21 @@ export class WorkflowControllerExternal {
   @common.HttpCode(200)
   @swagger.ApiForbiddenResponse({ type: errors.ForbiddenException })
   async createCollectionFlowUrl(
-    @common.Body() { expiry, workflowRuntimeDataId, endUserId }: CreateCollectionFlowUrlDto,
-    @CurrentProject() currentProjectId: TProjectId,
+    @common.Body()
+    { workflowRuntimeDataId }: CreateCollectionFlowUrlDto,
   ) {
-    const expiresAt = new Date(Date.now() + (expiry || 30) * 24 * 60 * 60 * 1000);
+    const result = await this.workflowTokenService.findFirstByWorkflowruntimeDataIdUnscoped(
+      workflowRuntimeDataId,
+    );
 
-    const { token } = await this.workflowTokenService.create(currentProjectId, {
-      workflowRuntimeDataId: workflowRuntimeDataId,
-      expiresAt,
-      endUserId,
-    });
+    if (!result) {
+      throw new NotFoundException(
+        `No WorkflowRuntimeDataId was found for ${JSON.stringify(workflowRuntimeDataId)}`,
+      );
+    }
 
     return {
-      token,
-      collectionFlowUrl: `${env.COLLECTION_FLOW_URL}?token=${token}`,
+      collectionFlowUrl: `${env.COLLECTION_FLOW_URL}?token=${result.token}`,
     };
   }
 
@@ -404,10 +402,16 @@ export class WorkflowControllerExternal {
   @common.HttpCode(200)
   @swagger.ApiForbiddenResponse({ type: errors.ForbiddenException })
   async createToken(
-    @common.Body() body: CreateCollectionFlowUrlDto,
+    @common.Body() { expiry, workflowRuntimeDataId, endUserId }: CreateTokenDto,
     @CurrentProject() currentProjectId: TProjectId,
   ) {
-    const { token } = await this.createCollectionFlowUrl(body, currentProjectId);
+    const expiresAt = new Date(Date.now() + (expiry || 30) * 24 * 60 * 60 * 1000);
+
+    const { token } = await this.workflowTokenService.create(currentProjectId, {
+      workflowRuntimeDataId: workflowRuntimeDataId,
+      expiresAt,
+      endUserId,
+    });
 
     return {
       token,