+++ apps/backoffice-v2/CHANGELOG.md
@@ -1,5 +1,69 @@
 # @ballerine/backoffice-v2
 
+## 0.7.72
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.5.47
+  - @ballerine/workflow-browser-sdk@0.6.68
+  - @ballerine/workflow-node-sdk@0.6.68
+  - @ballerine/react-pdf-toolkit@1.2.47
+
+## 0.7.71
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.53
+  - @ballerine/ui@0.5.46
+  - @ballerine/workflow-browser-sdk@0.6.67
+  - @ballerine/workflow-node-sdk@0.6.67
+  - @ballerine/react-pdf-toolkit@1.2.46
+
+## 0.7.70
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/blocks@0.2.27
+  - @ballerine/common@0.9.52
+  - @ballerine/react-pdf-toolkit@1.2.45
+  - @ballerine/ui@0.5.45
+  - @ballerine/workflow-browser-sdk@0.6.66
+  - @ballerine/workflow-node-sdk@0.6.66
+
+## 0.7.69
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.6.65
+  - @ballerine/workflow-node-sdk@0.6.65
+
+## 0.7.68
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.51
+  - @ballerine/workflow-browser-sdk@0.6.64
+  - @ballerine/workflow-node-sdk@0.6.64
+
+## 0.7.67
+
+### Patch Changes
+
+- Cump
+- Updated dependencies
+  - @ballerine/blocks@0.2.26
+  - @ballerine/common@0.9.50
+  - @ballerine/react-pdf-toolkit@1.2.44
+  - @ballerine/ui@0.5.44
+  - @ballerine/workflow-browser-sdk@0.6.63
+  - @ballerine/workflow-node-sdk@0.6.63
+
 ## 0.7.66
 
 ### Patch Changes