+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @ballerine/headless-example
 
+## 0.3.67
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.68
+
+## 0.3.66
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.53
+  - @ballerine/workflow-browser-sdk@0.6.67
+
+## 0.3.65
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.52
+  - @ballerine/workflow-browser-sdk@0.6.66
+
+## 0.3.64
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.6.65
+
+## 0.3.63
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.51
+  - @ballerine/workflow-browser-sdk@0.6.64
+
+## 0.3.62
+
+### Patch Changes
+
+- Cump
+- Updated dependencies
+  - @ballerine/common@0.9.50
+  - @ballerine/workflow-browser-sdk@0.6.63
+
 ## 0.3.61
 
 ### Patch Changes