+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/blocks
 
+## 0.2.27
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.52
+
+## 0.2.26
+
+### Patch Changes
+
+- Cump
+- Updated dependencies
+  - @ballerine/common@0.9.50
+
 ## 0.2.25
 
 ### Patch Changes