+++ packages/common/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ballerine/common
 
+## 0.9.53
+
+### Patch Changes
+
+- Created a non JMESPath sanctions plugin using JS
+
+## 0.9.52
+
+### Patch Changes
+
+- version bump
+
+## 0.9.51
+
+### Patch Changes
+
+- Updated aml schema
+
+## 0.9.50
+
+### Patch Changes
+
+- Cump
+
 ## 0.9.49
 
 ### Patch Changes